azure_search_rag/pyproject.toml
[project]
name = "azure-search-rag"
version = "1.0.0"
description = "Azure Search RAG implementation"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "langchain-openai>=0.0.3",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",
    "requests>=2.0.0",
    "tqdm>=4.0.0",
]
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]